Corrected issue with assist level factor higher than 255

+++ firmware/common/src/state.c
@@ -169,7 +169,7 @@ void rt_send_tx_package(uint8_t type) {
         ui8_usart1_tx_buffer[3] = (uint8_t) rt_vars.ui8_walk_assist_level_factor[((rt_vars.ui8_assist_level) - 1)];
         ui8_usart1_tx_buffer[4] = 0;
       } else if (rt_vars.ui8_assist_level) {
-        uint16_t ui16_temp = (uint8_t) rt_vars.ui16_assist_level_factor[((rt_vars.ui8_assist_level) - 1)];
+        uint16_t ui16_temp = rt_vars.ui16_assist_level_factor[((rt_vars.ui8_assist_level) - 1)];
         ui8_usart1_tx_buffer[3] = (uint8_t) (ui16_temp & 0xff);
         ui8_usart1_tx_buffer[4] = (uint8_t) (ui16_temp >> 8);
